some change

+++ zcc-amc/src/main/java/com/wensheng/zcc/amc/dao/mysql/mapper/AmcAssetMapper.java
@@ -12,7 +12,7 @@ public interface AmcAssetMapper {
 
     int deleteByPrimaryKey(Long id);
 
-    int insert(AmcAsset record);
+    Long insert(AmcAsset record);
 
     int insertSelective(AmcAsset record);
